Puts the navigateToCard function on the card image instead of the entire result
